# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AssetFlow API",
    "description": "AI-native family asset allocation advisor backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/profiles/": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get User Profile",
        "description": "Get current user's profile with asset analysis",
        "operationId": "get_user_profile_api_v1_profiles__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_UserProfileWithAnalysisResponse_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profiles"
        ],
        "summary": "Update User Profile",
        "description": "Update user profile with re-analysis of assets",
        "operationId": "update_user_profile_api_v1_profiles__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_UserProfileWithAnalysisResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Create User Profile",
        "description": "Create user profile with asset analysis",
        "operationId": "create_user_profile_api_v1_profiles__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_UserProfileWithAnalysisResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "profiles"
        ],
        "summary": "Delete User Profile",
        "description": "Delete user profile",
        "operationId": "delete_user_profile_api_v1_profiles__delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_dict_str__Any__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_UserProfileWithAnalysisResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserProfileWithAnalysisResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "APIResponse[UserProfileWithAnalysisResponse]"
      },
      "APIResponse_dict_str__Any__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "APIResponse[dict[str, Any]]"
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "INTERNAL_ERROR",
          "VALIDATION_ERROR",
          "NOT_FOUND",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_TOKEN",
          "TOKEN_EXPIRED",
          "INVALID_CREDENTIALS",
          "ASSET_NOT_FOUND",
          "ASSET_SAVE_ERROR",
          "INVALID_ASSET_DATA",
          "SEARCH_API_ERROR",
          "SEARCH_TIMEOUT",
          "PROPERTY_NOT_FOUND",
          "CHAT_SESSION_ERROR",
          "AI_SERVICE_ERROR",
          "DATABASE_ERROR",
          "CONNECTION_ERROR"
        ],
        "title": "ErrorCode",
        "description": "Standard error codes for API responses"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskLevel": {
        "type": "string",
        "enum": [
          "conservative",
          "moderate",
          "aggressive"
        ],
        "title": "RiskLevel"
      },
      "UserProfileCreate": {
        "properties": {
          "age_range": {
            "type": "string",
            "title": "Age Range"
          },
          "family_structure": {
            "type": "string",
            "title": "Family Structure"
          },
          "risk_preference": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "monthly_expense": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Expense"
          }
        },
        "type": "object",
        "required": [
          "age_range",
          "family_structure",
          "risk_preference"
        ],
        "title": "UserProfileCreate",
        "description": "User profile creation model"
      },
      "UserProfileUpdate": {
        "properties": {
          "age_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Range"
          },
          "family_structure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Structure"
          },
          "risk_preference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "monthly_expense": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Expense"
          }
        },
        "type": "object",
        "title": "UserProfileUpdate",
        "description": "User profile update model"
      },
      "UserProfileWithAnalysisResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "age_range": {
            "type": "string",
            "title": "Age Range"
          },
          "family_structure": {
            "type": "string",
            "title": "Family Structure"
          },
          "risk_preference": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "monthly_expense": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Expense"
          },
          "asset_analysis": {
            "additionalProperties": true,
            "type": "object",
            "title": "Asset Analysis"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "age_range",
          "family_structure",
          "risk_preference",
          "asset_analysis"
        ],
        "title": "UserProfileWithAnalysisResponse",
        "description": "User profile response with asset analysis"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}